+++ Nightmare_Prototype/Assets/Test/Scripts/StageGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb6fae8842e09564c8a8abb7c1ad0077
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
