+++ Nightmare_Prototype/Assets/Test/Scripts/StageViewer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3ae5680e518ee44faa14795a2b0109d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
